requirements.txt
pygame
pymodi
pymodi-plus
SpeechRecognition
numpy
pandas
opencv-python
tensorflow==2.15.0
matplotlib
jupyterlab
pyaudio
openpyxl